requirements/base.txt


# Bleeding edge Django
django==1.10.5

# Configuration
django-environ==0.4.1



# Forms
django-braces==1.11.0
django-crispy-forms==1.6.1

#multi-upload
django-multiupload==0.5.2

# Models
django-model-utils==2.6.1

# Images
Pillow==4.0.0

# For user registration, either via email or social
# Well-built with regular release cycles!
django-allauth==0.30.0


# Python-PostgreSQL Database Adapter
psycopg2==2.6.2

# Unicode slugification
awesome-slugify==1.6.5

# Time zones support
pytz==2016.10

# Redis support
django-redis==4.7.0
redis>=2.10.5

#countries list
django-countries-plus==1.1


# Your custom requirements go here
django-recaptcha==1.2.0
